Box2D core implementation
This is the initial commit for the box2d core functionality. There is
still some work that needs to be done with physics contact listeners,
but for now the physics system is working as intended.

NOTE: The previous physics system is still lingering in the code, and
must be removed at a later time.