Core BoxPhysicsListener Implemented
The core features of the BoxPhyiscsListener have been implemented. I
also created another class that will be responsible for creating physics
objects: BoxPhysicsFactory. This factory will also distinguish which
sprites are a part of which sprite "grouping" i.e friendy sprites, enemy
sprites, projectiles, etc. The reason for this is to allow box2d to
distinguish whether a collision between two objects should matter or not
(we MAY not want two bots to collide, or two projectiles).